add netty

+++ rpc/rpc-api/src/main/java/org/itourshare/api/service/UserService.java
@@ -4,6 +4,6 @@ import org.itourshare.api.entity.User;
 
 public interface UserService {
 
-    User getUser(Long id);
+    User getUser(Integer id);
 
 }
